Sanitize project search parameters earlier
This fixes a problem when e.g. maintainer is specified in projects
search form with mixed case and/or leading/trailing spaces. It's
intended to be normalized (stripped and lowercased) for all uses,
but is in fact only normalized when being passed to the database
query, but not to the code which selects packages belonging to the
specified maintainer to fill "Selected" column in the projects table.