Rework feed limitations (#182)
Events in atom feeds are limited by age to save clients from
refetching hundreds of old events again and again. This, however,
allows the feed to become empty if nothing happens for long enough
time, which was not intended.

So rework atom feeds to always return a few events even if these
are very old. Add all related knobs to the config (and split
from a project history knob).